+++ Realm/RLMObjectStore.h
@@ -61,6 +61,9 @@ RLMObjectBase *RLMCreateObjectInRealmWithValue(RLMRealm *realm, NSString *classN
                                                id _Nullable value, RLMUpdatePolicy updatePolicy)
 NS_RETURNS_RETAINED;
 
+// creates an asymmetric object and doesn't return
+void RLMCreateAsymmetricObjectInRealm(RLMRealm *realm, NSString *className, id value);
+
 //
 // Accessor Creation
 //